SREP-1919 - Deploy internal monitoring objects once both kube-apiserver reachable and vpce ready.
Slightly rework hcpReady() func to return bool and error, to more
clearly differentiate between apiserver not ready and error in check.
